+++ src/Forms/Fields/Concerns/HasRedactorToolbar.php
@@ -35,10 +35,10 @@ trait HasRedactorToolbar
             $this->redactorOptions = array_merge($this->redactorOptions, ['direction' => 'rtl']);
         }
 
-        return array_merge($this->defaultRedactorOptions(), $this->redactorOptions);
+        return array_merge($this->defaultRedactorOptions($locale), $this->redactorOptions);
     }
 
-    private function defaultRedactorOptions(): array
+    private function defaultRedactorOptions(?string $locale = null): array
     {
         return [
         ];