+++ src/Forms/UI/views/fields/text-with-redactor.blade.php
@@ -4,5 +4,5 @@
 ></div>
 
 <x-chief::form.input.prepend-append :prepend="$getPrepend($locale ?? null)" :append="$getAppend($locale ?? null)">
-    @include('chief-form::fields.html')
+    @include ('chief-form::fields.html')
 </x-chief::form.input.prepend-append>